Commit from JBoss Tools

+++ src/main/java/it/fperfetti/asos/botteghino/stub/Event.java
@@ -20,7 +20,7 @@ import javax.xml.bind.annotation.XmlType;
  *         &lt;element name="availability" type="{http://www.w3.org/2001/XMLSchema}int" minOccurs="0"/>
  *         &lt;element name="category" type="{http://fornitore-fabioperfetti.rhcloud.com/jboss-as-helloworld-ws/FornitoreService}category" minOccurs="0"/>
  *         &lt;element name="description" type="{http://www.w3.org/2001/XMLSchema}string" minOccurs="0"/>
- *         &lt;element name="id" type="{http://www.w3.org/2001/XMLSchema}int" minOccurs="0"/>
+ *         &lt;element name="id" type="{http://www.w3.org/2001/XMLSchema}long" minOccurs="0"/>
  *         &lt;element name="location" type="{http://www.w3.org/2001/XMLSchema}string" minOccurs="0"/>
  *         &lt;element name="price" type="{http://www.w3.org/2001/XMLSchema}double" minOccurs="0"/>
  *         &lt;element name="title" type="{http://www.w3.org/2001/XMLSchema}string" minOccurs="0"/>
@@ -49,7 +49,7 @@ public class Event {
     protected Integer availability;
     protected Category category;
     protected String description;
-    protected Integer id;
+    protected Long id;
     protected String location;
     protected Double price;
     protected String title;
@@ -155,10 +155,10 @@ public class Event {
      * 
      * @return
      *     possible object is
-     *     {@link Integer }
+     *     {@link Long }
      *     
      */
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
@@ -167,10 +167,10 @@ public class Event {
      * 
      * @param value
      *     allowed object is
-     *     {@link Integer }
+     *     {@link Long }
      *     
      */
-    public void setId(Integer value) {
+    public void setId(Long value) {
         this.id = value;
     }
